Traverser only interacts with raw_array_builder, so move them together
